backend_library/requirements.txt
fastapi[standard]>=0.128
uvicorn[standard]>=0.40
sqlalchemy[asyncio]>=2.0
asyncpg>=0.31
sentence-transformers>=5.2
openai>=2.17
httpx>=0.28
langgraph>=1.0
neo4j>=6.1
pydantic-settings>=2.12
numpy>=2.2
scikit-learn>=1.7
python-dotenv>=1.2
